requirements.txt
pinecone
langchain
langchain-core
langchain-groq
python-dotenv